Remove probe token mechanism in favor of manually signed tokens
