+++ README.md
@@ -193,8 +193,6 @@ Application Options:
       --user-id-path=                                       Dot notation path of a UserID for use with whitelist and X-Forwarded-User (default: email) [$USER_ID_PATH]
       --whitelist=                                          Only allow given UserID, comma separated, can be set multiple times [$WHITELIST]
       --port=                                               Port to listen on (default: 4181) [$PORT]
-      --probe-token=                                        Static probe token which is always passed [$PROBE_TOKEN]
-      --probe-token-user=                                   User authenticated with static probe token (default: probe) [$PROBE_TOKEN_USER]
       --rule.<name>.<param>=                                Rule definitions, param can be: "action", "rule" or "provider"
       --trusted-ip-address=                                 List of trusted IP addresses or IP networks (in CIDR notation) that are considered authenticated [$TRUSTED_IP_ADDRESS]
