+++ internal/config.go
@@ -43,8 +43,6 @@ type Config struct {
 	UserPath               string               `long:"user-id-path" env:"USER_ID_PATH" default:"email" description:"Dot notation path of a UserID for use with whitelist and X-Forwarded-User"`
 	Whitelist              CommaSeparatedList   `long:"whitelist" env:"WHITELIST" env-delim:"," description:"Only allow given UserID, comma separated, can be set multiple times"`
 	Port                   int                  `long:"port" env:"PORT" default:"4181" description:"Port to listen on"`
-	ProbeToken             CommaSeparatedList   `long:"probe-token" env:"PROBE_TOKEN" env-delim:"," description:"Static probe token which is always passed"`
-	ProbeTokenUser         string               `long:"probe-token-user" env:"PROBE_TOKEN_USER" default:"probe" description:"User authenticated with static probe token"`
 
 	Providers provider.Providers `group:"providers" namespace:"providers" env-namespace:"PROVIDERS"`
 	Rules     map[string]*Rule   `long:"rule.<name>.<param>" description:"Rule definitions, param can be: \"action\", \"rule\" or \"provider\""`
