+++ internal/token_test.go
@@ -3,78 +3,71 @@ package tfa
 import (
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
 )
 
-const (
-	testSecret = "SLUGoz2n44q2Wh-dQRXe-IheSpjfcBzy"
-	testUser   = "test-user"
-)
+func Test_verifyToken(t *testing.T) {
+	t.Run("should not pass empty with default", func(t *testing.T) {
+		assert := assert.New(t)
+		config, _ = NewConfig([]string{""})
 
-func BenchmarkVerifyToken_Original(b *testing.B) {
-	config.Secret = []byte(testSecret)
-
-	expiry := time.Now().Unix() + 100
-	tok := token(testUser, expiry)
-
-	b.StartTimer()
-	for i := 0; i < b.N; i++ {
-		verifiedUser, err := verifyToken(tok)
-		if err != nil {
-			b.Fatal(err)
+		_, err := verifyToken("")
+		if assert.Error(err) {
+			assert.Equal("Invalid cookie format", err.Error())
 		}
-		if verifiedUser != testUser {
-			b.Fatal("verified user is wrong")
+	})
+
+	t.Run("should require 3 parts", func(t *testing.T) {
+		assert := assert.New(t)
+		config, _ = NewConfig([]string{""})
+
+		_, err := verifyToken("")
+		if assert.Error(err) {
+			assert.Equal("Invalid cookie format", err.Error())
 		}
-	}
-}
-
-// Below contains code for comparison with JWT-based token implementation.
-// They are commented out because they depend on github.com/golang-jwt/jwt/v5 library.
-// JWT-based impl. is about 50~100 times slower than the original HMAC256-based implementation.
-/*
-func GenerateECDSAKey() (privRaw []byte, pubRaw []byte) {
-	priv, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	ecder, _ := x509.MarshalECPrivateKey(priv)
-	ecderpub, _ := x509.MarshalPKIXPublicKey(&priv.PublicKey)
-	return pem.EncodeToMemory(&pem.Block{Type: "EC PRIVATE KEY", Bytes: ecder}), pem.EncodeToMemory(&pem.Block{Type: "PUBLIC KEY", Bytes: ecderpub})
-}
-
-var privRaw []byte
-var privKey *ecdsa.PrivateKey
-
-func init() {
-	privRaw, _ = GenerateECDSAKey()
-	privKey = lo.Must(jwt.ParseECPrivateKeyFromPEM(privRaw))
-}
-
-func signJWT(user string, expiry int64) string {
-	now := time.Now()
-	claims := jwt.MapClaims{
-		"iat": now.Unix(),
-		"exp": expiry,
-		"sub": user,
-	}
-	return lo.Must(jwt.NewWithClaims(jwt.SigningMethodES256, claims).SignedString(privKey))
-}
-
-func verifyJWT(token string) (string, error) {
-	tok := lo.Must(jwt.Parse(token, func(token *jwt.Token) (any, error) { return &privKey.PublicKey, nil }))
-	return tok.Claims.GetSubject()
-}
-
-func BenchmarkVerifyToken_JWT(b *testing.B) {
-	expiry := time.Now().Unix() + 100
-	tok := signJWT(testUser, expiry)
-
-	b.StartTimer()
-	for i := 0; i < b.N; i++ {
-		verifiedUser, err := verifyJWT(tok)
-		if err != nil {
-			b.Fatal(err)
+		_, err = verifyToken("1|2")
+		if assert.Error(err) {
+			assert.Equal("Invalid cookie format", err.Error())
 		}
-		if verifiedUser != testUser {
-			b.Fatal("verified user is wrong")
+		_, err = verifyToken("1|2|3|4")
+		if assert.Error(err) {
+			assert.Equal("Invalid cookie format", err.Error())
 		}
-	}
+	})
+
+	t.Run("should catch invalid mac", func(t *testing.T) {
+		assert := assert.New(t)
+		config, _ = NewConfig([]string{""})
+
+		_, err := verifyToken("MQ==|2|3")
+		if assert.Error(err) {
+			assert.Equal(ErrInvalidSignature, err)
+		}
+	})
+
+	t.Run("should catch expired", func(t *testing.T) {
+		assert := assert.New(t)
+		config, _ = NewConfig([]string{""})
+
+		expiry := time.Now().Add(-time.Second).Unix()
+		tok := token("test@test.com", expiry)
+
+		_, err := verifyToken(tok)
+		if assert.Error(err) {
+			assert.Equal("Cookie has expired", err.Error())
+		}
+	})
+
+	t.Run("should accept valid cookie", func(t *testing.T) {
+		assert := assert.New(t)
+		config, _ = NewConfig([]string{""})
+
+		expiry := time.Now().Add(10 * time.Second).Unix()
+		tok := token("test@test.com", expiry)
+
+		email, err := verifyToken(tok)
+		assert.Nil(err, "valid request should not return an error")
+		assert.Equal("test@test.com", email, "valid request should return user email")
+	})
 }
-*/
